requirements.txt
elevenlabs 
pyaudio 
python-dotenv